backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
PyPDF2==3.0.1
transformers==4.35.0
torch==2.1.0
google-cloud-firestore==2.13.0
python-multipart==0.0.6

